+++ Src/Vector2D.cpp
@@ -6,7 +6,7 @@ Vector2D::Vector2D()
 	y = 0.0f;
 }
 
-Vector2D::Vector2D(float _x, float _y)
+Vector2D::Vector2D(int _x, int _y)
 {
 	x = _x;
 	y = _y;
